+++ R/create_codemeta.R
@@ -2,7 +2,7 @@
 #' create_codemeta
 #'
 #' create a codemeta list object in R for further manipulation. Similar
-#' to \code{\link{write_codemeta}}, but returns an R list object rather
+#' to [write_codemeta()], but returns an R list object rather
 #' than writing directly to a file.  See examples.
 #'
 #' @inheritParams write_codemeta
@@ -27,7 +27,6 @@ create_codemeta <- function(
 
   ## looks like we got a package name/path or Description file
   if (is.character(pkg)) {
-
     root <- get_root_path(pkg)
 
     # Set string constants
@@ -41,13 +40,11 @@ create_codemeta <- function(
 
       ## no cm, no existing codemeta.json found, start fresh
     } else {
-
       new_codemeta()
     }
 
     ## we got an existing codemeta object as pkg
   } else if (is.list(pkg)) {
-
     cm <- pkg
 
     ## root should be set already, we might check that root has a DESCRIPTION,
@@ -56,12 +53,11 @@ create_codemeta <- function(
   }
 
   if (verbose) {
-
     root <- get_root_path(pkg)
-    opinions <- give_opinions(root)
 
-    if (! is.null(opinions)) {
+    opinions <- give_opinions(root, verbose)
 
+    if (!is.null(opinions)) {
       message(
         "Some elements could be improved, see our opinions via give_opinions('",
         root, "')"
@@ -70,48 +66,58 @@ create_codemeta <- function(
   }
 
   ## get information from DESCRIPTION
-  cm <- codemeta_description(file.path(root, "DESCRIPTION"), id = id, cm)
+  cm <- codemeta_description(file.path(root, "DESCRIPTION"), id = id, cm,
+                             verbose = verbose)
 
   ## Guess these only if not set in current codemeta
-  if ((is.null(cm$codeRepository) && force_update)) {
+  # try to identify a code repo
 
-    cm$codeRepository <- guess_github(root)
+  more_work_cr <- function(codeRepository) {
+    if (is.null(codeRepository)) {
+      return(TRUE)
+    }
+
+    !urltools::domain(codeRepository) %in% source_code_domains()
+  }
+
+  if (more_work_cr(cm$codeRepository)) {
+
+    if (!is.null(guess_github(root)) && force_update) {
+      cm$relatedLink <- cm$codeRepository
+      cm$codeRepository <- guess_github(root)
+    }
+
   }
 
   if ((is.null(cm$releaseNotes) || force_update)) {
-
-    cm$releaseNotes <- guess_releaseNotes(root)
+    cm$releaseNotes <- guess_releaseNotes(root, cm)
   }
 
   if ((is.null(cm$readme) || force_update)) {
-
-    cm$readme <- guess_readme(root)$readme_url
+    cm$readme <- guess_readme(root, verbose, cm)$readme_url
   }
 
   if (use_filesize) {
-
     if ((is.null(cm$fileSize) || force_update)) {
-
       cm$fileSize <- guess_fileSize(root)
     }
   }
 
   # and if there's a readme
-  readme <- guess_readme(root)$readme_path
+  readme <- guess_readme(root, verbose, cm)$readme_path
 
-  if (! is.null(readme) && force_update) {
-
+  if (!is.null(readme) && force_update) {
     cm <- codemeta_readme(readme, codemeta = cm)
   }
 
   ## If code repo is GitHub
-  if (grepl("github.com\\/.*\\/.*", cm$codeRepository)) {
-
-    cm <- add_github_topics(cm)
+  if (!is.null(cm$codeRepository) && urltools::domain(cm$codeRepository) %in%
+    github_domains()) {
+    cm <- add_github_topics(cm, verbose)
   }
 
   ## Citation metadata
-  if (is.character(pkg)) {  ## Doesn't apply if pkg is a list (codemeta object)
+  if (is.character(pkg)) { ## Doesn't apply if pkg is a list (codemeta object)
 
     cm$citation <- guess_citation(pkg)
 
@@ -120,8 +126,7 @@ create_codemeta <- function(
 
     ## citations need schema.org context!
     ## see https://github.com/codemeta/codemeta/issues/155
-    if (! any(grepl(url_schema, cm$`@context`))) {
-
+    if (!any(grepl(url_schema, cm$`@context`))) {
       cm$`@context` <- c(cm$`@context`, url_schema)
     }
   }
@@ -130,29 +135,24 @@ create_codemeta <- function(
   # Priority is given to the README
   # alternatively to installed packages
 
-  provider <- guess_provider(cm$identifier)
+  provider <- guess_provider(cm$identifier, verbose)
 
-  if (! is.null(provider)) {
+  if (!is.null(provider)) {
+    readme <- guess_readme_path(root)
 
-    readme <- guess_readme(root)$readme_path
-
-    if (! is.null(readme)) {
-
+    if (!is.null(readme)) {
       badges <- extract_badges(readme)
 
-      if (! is.null(provider) &&
-          whether_provider_badge(badges, provider$name)) {
-
+      if (!is.null(provider) &&
+        whether_provider_badge(badges, provider$name)) {
         cm <- set_relatedLink_1(cm, provider)
       }
-    } else if (cm$identifier %in% installed_package_names()) {
-
+    } else if (is_installed(cm$identifier)) {
       pkg_info <- sessioninfo::package_info(cm$identifier)
       pkg_info <- pkg_info[pkg_info$package == cm$identifier, ]
       provider_name <- pkg_info$source
 
       if (cm$version == pkg_info$ondiskversion) {
-
         cm <- set_relatedLink_2(cm, provider_name)
       }
     }
@@ -164,15 +164,11 @@ create_codemeta <- function(
 
 # set_relatedLink_1 ------------------------------------------------------------
 set_relatedLink_1 <- function(codemeta, provider) {
-
   if (provider$name == "Comprehensive R Archive Network (CRAN)") {
-
     codemeta$relatedLink <- unique(c(
       codemeta$relatedLink, get_url_cran_package(codemeta$identifier)
     ))
-
   } else if (provider$name == "BioConductor") {
-
     codemeta$relatedLink <- unique(c(
       codemeta$relatedLink, get_url_bioconductor_package(codemeta$identifier)
     ))
@@ -183,24 +179,18 @@ set_relatedLink_1 <- function(codemeta, provider) {
 
 # set_relatedLink_2 ------------------------------------------------------------
 set_relatedLink_2 <- function(codemeta, provider_name) {
-
   if (grepl("CRAN", provider_name)) {
-
     codemeta$relatedLink <- unique(c(
       codemeta$relatedLink, get_url_cran_package(codemeta$identifier)
     ))
-
   } else if (grepl("Bioconductor", provider_name)) {
-
     codemeta$relatedLink <- unique(c(
       codemeta$relatedLink, get_url_bioconductor_package(codemeta$identifier)
     ))
-
   } else if (grepl("Github", provider_name)) {
 
     # if GitHub try to build the URL to commit or to repo in general
     if (grepl("@", provider_name)) {
-
       codemeta$relatedLink <- unique(c(
         codemeta$relatedLink, get_url_github_package(provider_name)
       ))