+++ R/give_opinions.R
@@ -3,11 +3,12 @@
 #' Function giving opinions about a package
 #'
 #' @param pkg_path Path to the package root
+#' @param verbose Whether to print message related to internet download progress.
 #'
 #' @return A data.frame of opinions
 #' @export
 #'
-give_opinions <- function(pkg_path = getwd()) {
+give_opinions <- function(pkg_path = getwd(), verbose = FALSE) {
 
   # set the path to the description file
   description_file <- file.path(pkg_path, "DESCRIPTION")
@@ -22,7 +23,7 @@ give_opinions <- function(pkg_path = getwd()) {
     give_opinions_desc(description = description),
 
     # opinions about README
-    try_to_give_opinions_readme(description_file)
+    try_to_give_opinions_readme(description_file, verbose)
   ))
 
   if (! is.null(fixmes)) {
@@ -37,12 +38,12 @@ give_opinions <- function(pkg_path = getwd()) {
 
 #' Read Description from File If Not Given
 #'
-#' @param description object to be checked for \code{NULL} and to be returned if
-#'   not being \code{NULL}.
-#' @param description_file path to \code{DESCRIPTION} file to be read with
-#'   \code{\link[desc]{desc}} if \code{description} is \code{NULL}.
-#' @return \code{description} if not \code{NULL} or the content of
-#'   \code{description_file} read with \code{\link[desc]{desc}}.
+#' @param description object to be checked for `NULL` and to be returned if
+#'   not being `NULL`.
+#' @param description_file path to `DESCRIPTION` file to be read with
+#'   [desc::desc()] if `description` is `NULL`.
+#' @return `description` if not `NULL` or the content of
+#'   `description_file` read with [desc::desc()].
 #' @noRd
 read_description_if_null <- function(description, description_file) {
 
@@ -61,14 +62,14 @@ read_description_if_null <- function(description, description_file) {
 #' Give Opinions about DESCRIPTION File
 #'
 #' You may either pass the path to a DESCRIPTION file or a description object
-#' as read with \code{\link[desc]{desc}}.
+#' as read with [desc::desc()].
 #'
 #' @param description_file path do a DESCRIPTION file. Will not be considered if
-#'   a description object is given in \code{description}.
+#'   a description object is given in `description`.
 #' @param description Description object as read with
-#'   \code{\link[desc]{desc}}. If not provided, the path to a DESCRIPTION file
-#'   must be given in \code{description_file}.
-#' @return \code{tibble} with columns \code{where}, \code{fixme} or \code{NULL}
+#'   [desc::desc()]. If not provided, the path to a DESCRIPTION file
+#'   must be given in `description_file`.
+#' @return `tibble` with columns `where`, `fixme` or `NULL`
 #'   if there are no opionions about the DESCRIPTION file.
 #' @noRd
 give_opinions_desc <- function(description_file, description = NULL) {
@@ -134,22 +135,23 @@ fixmes_as_tibble_or_message <- function(fixmes, where, message_id = NULL) {
 }
 
 # try_to_give_opinions_readme --------------------------------------------------
-try_to_give_opinions_readme <- function(description_file) {
+try_to_give_opinions_readme <- function(description_file, verbose = FALSE) {
 
-  readme_path <- guess_readme(dirname(description_file))$readme_path
+  readme_path <- guess_readme_path(dirname(description_file))
 
   if (is.null(readme_path)) {
 
     return(NULL)
   }
 
-  desc_info <- codemeta_description(description_file)
-
-  give_opinions_readme(readme_path, pkg_name = desc_info$identifier)
+  give_opinions_readme(readme_path,
+                       pkg_name = desc::desc_get_field("Package",
+                                                       file = description_file),
+                       verbose)
 }
 
 # give_opinions_readme ---------------------------------------------------------
-give_opinions_readme <- function(readme_path, pkg_name) {
+give_opinions_readme <- function(readme_path, pkg_name, verbose = FALSE) {
 
   # start with an empty vector of "fixme" messages
   fixmes <- character()
@@ -161,7 +163,7 @@ give_opinions_readme <- function(readme_path, pkg_name) {
   }
 
   # provider
-  provider <- guess_provider(pkg_name)
+  provider <- guess_provider(pkg_name, verbose)
 
   if (has_provider_but_no_badge(provider, readme_path)) {