+++ R/parse_citation.R
@@ -1,11 +1,10 @@
 ## map a "citation" or "bibentry" R object into schema.org
 # bib <- citation(pkg)
 
-#' @importFrom stringi stri_trans_general
 parse_citation <- function(bib) {
 
   type <- bib$bibtype %>%
-    stringi::stri_trans_general(id = "Title") %>%
+    tools::toTitleCase() %>%
     bibentry_to_schema_field()
 
   author <- parse_people(bib$author, new_codemeta())$author
@@ -61,7 +60,7 @@ init_citation <- function(type, author, doi, id, bib)
     "identifier" = doi,
     "url" = bib$url,
     "description" = bib$note,
-    "paginiation" = bib$pages,
+    "pagination" = bib$pages,
     "@id" = id,   # may be NULL and will be removed by drop_null()
     "sameAs" = id # same same
   ))
@@ -69,7 +68,6 @@ init_citation <- function(type, author, doi, id, bib)
 
 # to_url_doi_or_null -----------------------------------------------------------
 
-#' @importFrom stringi stri_startswith
 to_url_doi_or_null <- function(doi) {
 
   # Return NULL if doi is NULL itself
@@ -79,7 +77,7 @@ to_url_doi_or_null <- function(doi) {
   }
 
   # Return doi if it already looks like an URL of doi.org
-  if (stringi::stri_startswith(doi, coll = get_url_doi())) {
+  if (grepl(paste0("^", get_url_doi()), doi)) {
 
     return(doi)
   }
@@ -129,7 +127,7 @@ guess_citation <- function(pkg) {
 
   citation_file_exists <- file.exists(citation_file)
 
-  package_is_installed <- pkg %in% installed_package_names()
+  package_is_installed <- is_installed(pkg)
 
   # Return NULL if there is no citation file and if pkg is not installed
   if (! citation_file_exists && ! package_is_installed) {