+++ R/utils.R
@@ -7,9 +7,9 @@ drop_null <- function(x) {
 # get_root_path ----------------------------------------------------------------
 get_root_path <- function(pkg) {
 
-  if (pkg %in% installed_package_names()) {
+  if (is_installed(pkg)) {
 
-    package_file(pkg, ".")
+    find.package(pkg)
 
   } else if (is_package(file.path(pkg))) {
 
@@ -32,11 +32,15 @@ package_file <- function(pkg, ...) {
   system.file(..., package = pkg)
 }
 
-# installed_package_names: names of installed packages -------------------------
-#' @importFrom utils installed.packages
-installed_package_names <- function() {
+# is_installed: is the package installed -------------------------
+is_installed <- function(pkg) {
 
-  row.names(installed.packages())
+  length(
+    find.package(
+      package = pkg,
+      quiet = TRUE
+      )
+    ) > 0
 }
 
 # get_file ---------------------------------------------------------------------
@@ -66,9 +70,10 @@ is_IRI <- function(string) {
 # uses_git ---------------------------------------------------------------------
 # from usethis cf https://github.com/r-lib/usethis/blob/2abb0422a97808cc573fa5900a8efcfed4c2d5b4/R/git.R#L68
 # this is GPL-3 code
+# now with gert not git2r
 uses_git <- function(path = usethis::proj_get()) {
 
-  ! is.null(git2r::discover_repository(path))
+  !is.null(tryCatch(gert::git_find(path), error = function(e){NULL}))
 }
 
 # from usethis cf https://github.com/r-lib/usethis/blob/4fb556788d2588facaaa8560242d2c83f2261d6e/R/helpers.R#L55
@@ -117,6 +122,11 @@ get_url_status_code <- function(url) {
 # check_urls -------------------------------------------------------------------
 check_urls <- function(urls) {
 
+  if (!pingr::is_online()) {
+
+    return("")
+  }
+
   messages <- do.call(rbind, lapply(urls, get_url_status_code))
 
   failed <- (messages$message != "All good")
@@ -151,26 +161,25 @@ is_package <- function(path) {
   all(c("DESCRIPTION", "NAMESPACE", "man", "R") %in% dir(path))
 }
 
-# set_element_if_null ----------------------------------------------------------
-set_element_if_null <- function(x, element, value) {
+# set_element ----------------------------------------------
+set_element <- function(x, element, value) {
 
   stopifnot(is.list(x))
 
-  if (is.null(x[[element]])) {
-
-    x[[element]] <- value
-  }
+  x[[element]] <- value
 
   x
+
 }
+
 
 # fails ------------------------------------------------------------------------
 #' Does the Evaluation of an Expression Fail?
 #'
-#' @param expr expression to be evaluated within \code{try(\dots)}
-#' @param silent passed to \code{\link{try}}, see there.
-#' @return \code{TRUE} if evaluating \code{expr} failed and \code{FALSE} if
-#'   the evalutation of \code{expr} succeeded.
+#' @param expr expression to be evaluated within `try(\dots)`
+#' @param silent passed to [try()], see there.
+#' @return `TRUE` if evaluating `expr` failed and `FALSE` if
+#'   the evalutation of `expr` succeeded.
 #' @noRd
 fails <- function(expr, silent = TRUE) {
 
@@ -187,8 +196,8 @@ example_file <- function(...) {
 #' Check for Class "json" or Character
 #'
 #' @param x object to be checked for its class and mode
-#' @return \code{TRUE} if \code{x} inherits from "json" or is of mode character,
-#'   otherwise \code{FALSE}
+#' @return `TRUE` if `x` inherits from "json" or is of mode character,
+#'   otherwise `FALSE`
 #' @noRd
 is_json_or_character <- function(x) {
 
@@ -201,7 +210,7 @@ is_json_or_character <- function(x) {
 #'
 #' @param condition expression to be evaluated
 #' @param x object to be converted to JSON
-#' @param \dots further arguments passed to \code{\link[jsonlite]{toJSON}}
+#' @param \dots further arguments passed to [jsonlite::toJSON()]
 #' @importFrom jsonlite toJSON
 #' @noRd
 to_json_if <- function(condition, x, ...) {
@@ -214,9 +223,9 @@ to_json_if <- function(condition, x, ...) {
 #' Convert from JSON if Condition is Met
 #'
 #' @param condition expression to be evaluated
-#' @param x object passed to \code{\link[jsonlite]{fromJSON}} if
-#'   \code{condition} is met
-#' @param \dots further arguments passed to \code{\link[jsonlite]{fromJSON}}
+#' @param x object passed to [jsonlite::fromJSON()] if
+#'   `condition` is met
+#' @param \dots further arguments passed to [jsonlite::fromJSON()]
 #' @importFrom jsonlite fromJSON
 #' @noRd
 from_json_if <- function(condition, x, ...) {
@@ -229,9 +238,9 @@ from_json_if <- function(condition, x, ...) {
 #' Call Function if Condition is Met
 #'
 #' @param condition expression to be evaluated
-#' @param FUN function to be called if \code{condition} is met
-#' @param x first argument to be passed to \code{FUN} or not
-#' @param \dots further arguments passed to \code{FUN}
+#' @param FUN function to be called if `condition` is met
+#' @param x first argument to be passed to `FUN` or not
+#' @param \dots further arguments passed to `FUN`
 #' @noRd
 call_if <- function(condition, x, FUN, ...) {
 
@@ -243,4 +252,9 @@ call_if <- function(condition, x, FUN, ...) {
 
     x
   }
+}
+
+# bind df -----------------------
+bind_df <- function(dfs) {
+  do.call("rbind", dfs)
 }