+++ R/write_codemeta.R
@@ -4,7 +4,10 @@
 #' basically a wrapper around create_codemeta() to both create the codemeta
 #' object and write it out to a JSON-LD-formatted file in one command. It can
 #' also be used simply to write out to JSON-LD any existing object created with
-#' create_codemeta().
+#' `create_codemeta()`.
+#'
+#' @includeRmd man/rmdhunks/whybother.Rmd
+#' @includeRmd man/rmdhunks/uptodate.Rmd
 #'
 #' @param pkg package path to package root, or package name, or description file
 #'   (character), or a codemeta object (list)
@@ -13,26 +16,28 @@
 #'   root. Default guess is current dir.
 #' @param id identifier for the package, e.g. a DOI (or other resolvable URL)
 #' @param use_filesize whether to try to estimating and adding a filesize by using
-#'   \code{base::files.ize()}. Files in \code{.Rbuildignore} are ignored.
+#'   `base::file.size()`. Files in `.Rbuildignore` are ignored.
 #' @param force_update Update guessed fields even if they are defined in an
 #'   existing codemeta.json file
-#' @param use_git_hook Whether to create a pre-commit hook requiring
-#'   codemeta.json to be updated when DESCRIPTION is changed. Default is \code{FALSE} to avoid
-#'   an unwanted alteration of the user's git environment.
+#' @param use_git_hook Deprecated argument.
 #' @param verbose Whether to print messages indicating opinions e.g. when
-#'   DESCRIPTION has no URL. See \code{\link{give_opinions}}.
+#'   DESCRIPTION has no URL. -- See \code{\link{give_opinions}};
+#' and indicating the progress of internet downloads.
+#' @param write_minimeta whether to also create the file schemaorg.json that
+#' corresponds to the metadata Google would validate, to be inserted to a
+#' webpage for SEO. It is saved as "inst/schemaorg.json" alongside `path` (by
+#' default, "codemeta.json").
 #' @param ...  additional arguments to \code{\link{write_json}}
-#' @details If pkg is a codemeta object, the function will attempt to update any
+#' @section Technical details:
+#'  If pkg is a codemeta object, the function will attempt to update any
 #'   fields it can guess (i.e. from the DESCRIPTION file), overwriting any
 #'   existing data in that block. In this case, the package root directory
 #'   should be the current working directory.
 #'
 #' When creating and writing a codemeta.json for the first time, the function
-#' adds "codemeta.json" to .Rbuildignore and, if the project uses git, adds a
-#' pre-commit hook ensuring that if DESCRIPTION changes, the codemeta.json will
-#' be updated as well unless the DESCRIPTION change is committed with 'git
-#' commit --no-verify'.
-#' @return writes out the codemeta.json file
+#' adds "codemeta.json" to .Rbuildignore.
+#' @return writes out the codemeta.json file, and schemaorg.json if `write_codemeta`
+#' is `TRUE`.
 #' @export
 #'
 #' @examples
@@ -40,10 +45,14 @@
 #' write_codemeta("codemetar", path = "example_codemetar_codemeta.json")
 #' }
 write_codemeta <- function(
-  pkg = ".", path = "codemeta.json", root = ".", id = NULL, use_filesize = FALSE,
-  force_update = getOption("codemeta_force_update", TRUE), use_git_hook = FALSE,
-  verbose = TRUE, ...
+  pkg = ".", path = "codemeta.json", root = ".", id = NULL, use_filesize = TRUE,
+  force_update = getOption("codemeta_force_update", TRUE), use_git_hook = NULL,
+  verbose = TRUE, write_minimeta = FALSE, ...
 ) {
+
+  if (!missing(use_git_hook)) {
+    warning("The use_git_hook argument is deprecated and ignored.")
+  }
 
   codemeta_json <- "codemeta.json"
 
@@ -55,24 +64,22 @@ write_codemeta <- function(
   if (in_package && path == codemeta_json) {
 
     usethis::use_build_ignore(codemeta_json)
-
-    # Add the git pre-commit hook
-    # https://github.com/r-lib/usethis/blob/master/inst/templates/
-    #   readme-rmd-pre-commit.sh#L1
-    # this is GPL-3 code
-    no_codemeta_json <- ! file.exists(file.path(pkg, codemeta_json))
-
-    if (uses_git() && use_git_hook && no_codemeta_json) {
-
-      message(get_message("adding_hook")) # nolint
-
-      usethis::use_git_hook("pre-commit", render_template(
-        "description-codemetajson-pre-commit.sh"
-      ))
-    }
   }
-
   # Create or update codemeta and save to disk
   create_codemeta(pkg = pkg, root = root, use_filesize = use_filesize) %>%
     jsonlite::write_json(path, pretty = TRUE, auto_unbox = TRUE, ...)
+
+  # Create minimeta and save to disk
+  if (write_minimeta) {
+    if (!requireNamespace("jsonld", quietly = TRUE)) {
+      stop("Package jsonld required. Please install before re-trying.")
+    }
+    schemaorg <- system.file("schema", "schemaorg.json",
+                             package="codemetar")
+   jsonld::jsonld_frame("codemeta.json", schemaorg) %>%
+     jsonld::jsonld_compact('{"@context": "https://schema.org"}') %>%
+     writeLines(file.path(dirname(path), "inst", "schemaorg.json"))
+
+
+  }
 }