+++ tests/testthat/test-utils.R
@@ -6,8 +6,6 @@ testthat::test_that("get_root_path is covered", {
   testthat::expect_true(file.exists(x))
   x <- get_root_path("not_a_package")
   testthat::expect_equal(x, "not_a_package")
-
-  ## test
 })
 
 testthat::test_that("example_file() works", {